fix in qr layout. Has separated ButtonDrawer into a separate file
