+++ lib/ui/shared/cup_layout.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:kofe_vkusno/ui/shared/widgets/app_bar_custom/app_bar_custom.dart';
-import 'package:kofe_vkusno/ui/shared/widgets/coffee_drawer.dart';
+import 'package:kofe_vkusno/ui/shared/widgets/coffee_drawer/coffee_drawer.dart';
 import 'package:kofe_vkusno/ui/shared/widgets/cup_layout_widgets/circle_button.dart';
 import 'package:kofe_vkusno/ui/shared/widgets/cup_layout_widgets/cup_indicator.dart';
