+++ lib/ui/shared/main_layout.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:kofe_vkusno/ui/shared/widgets/coffee_drawer.dart';
+import 'package:kofe_vkusno/ui/shared/widgets/coffee_drawer/coffee_drawer.dart';
 
 import 'widgets/app_bar_custom/app_bar_custom.dart';
 
@@ -15,6 +15,7 @@ class MainLayout extends StatelessWidget{
       body: Padding(
         padding:  const EdgeInsets.fromLTRB(8, 40, 8, 26),
         child: Column(
+
         ),
       ),
     );
